bootstraprag/templates/qdrant/semantic_routing/requirements.txt
semantic-router[qdrant]
python-dotenv==1.0.1